Refactor structure; add healthcheck route
