+++ server2/main.go
@@ -4,54 +4,14 @@ import (
 	"errors"
 	"flag"
 	"fmt"
-	"time"
 
+	"github.com/aspenjames/library-app/server/middlewares"
+	"github.com/aspenjames/library-app/server/routes"
 	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/middleware/compress"
-	"github.com/gofiber/fiber/v2/middleware/etag"
-	"github.com/gofiber/fiber/v2/middleware/limiter"
-	"github.com/gofiber/fiber/v2/middleware/logger"
-	"github.com/gofiber/fiber/v2/middleware/monitor"
-	"github.com/gofiber/fiber/v2/middleware/recover"
-	"github.com/gofiber/fiber/v2/middleware/requestid"
-	"github.com/gofiber/fiber/v2/middleware/timeout"
 )
 
 // Port on which to run the API server.
 var port int
-
-// Add middlewares:
-//   - Compress:  compress responses
-//   - ETag:      caching header
-//   - Limiter:   rate limiter
-//   - Logger:    logging middleware
-//   - Monitor:   expose common metrics
-//   - Recover:   handle panics with app.ErrorHandler
-//   - RequestID: add unique X-Request-ID header
-func setupMiddlewares(app *fiber.App) *fiber.App {
-	app.Use(recover.New())
-	app.Use(limiter.New(limiter.Config{
-		Max:        5,
-		Expiration: 10 * time.Second,
-	}))
-	app.Use(logger.New(logger.Config{
-		TimeZone: "UTC",
-	}))
-	app.Use(compress.New())
-	app.Use(etag.New())
-	app.Use(requestid.New())
-
-	// Setup metrics endpoint
-	app.Get("/metrics", monitor.New(monitor.Config{
-		Title: "Library-app server metrics",
-	}))
-	return app
-}
-
-// Wrap route handlers in timeout
-func timeoutWrapper(h fiber.Handler) fiber.Handler {
-	return timeout.New(h, 5*time.Second)
-}
 
 func main() {
 	flag.IntVar(&port, "p", 8080, "Port on which to listen")
@@ -71,11 +31,8 @@ func main() {
 			})
 		},
 	})
-	setupMiddlewares(app)
-	app.Get("/", timeoutWrapper(func(c *fiber.Ctx) error {
-		return c.SendString("Hello, World!")
-	}))
-
+	middlewares.Setup(app)
+	routes.Setup(app)
 	addr := fmt.Sprintf(":%d", port)
 	app.Listen(addr)
 }
